requirements.txt
torch>=1.9.0
numpy
Pillow
gdown
gradio
opencv-python
scikit-learn==0.22
dlib
lpips
tqdm
wandb
